+++ client/components/Nav.tsx
@@ -13,8 +13,14 @@ function Nav() {
             nook
           </Link>
         </h1>
-
-        <p>add project</p>
+        <div className="navItems_right">
+          <Link to="/elements" className="elements link">
+            <p className="navlink">my elements</p>
+          </Link>
+          <Link to="#" className="addProjectNav link">
+            <p className="navlink">add project</p>
+          </Link>
+        </div>
       </nav>
     </>
   )
